Split program into multiple "modules"
The program is now split into multiple modules, to make collaboration easier.

Added combatModule, fileUtils, movementModule and navigationModule files.
